requirements.txt
anthropic==0.3.1
pandas>=1.5.3
requests>=2.28.0
python-dotenv>=1.0.0
termcolor>=2.3.0
streamlit>=1.26.0
matplotlib>=3.7.0
plotly>=5.15.0
numpy>=1.25.0
scipy>=1.11.0
ujson>=5.8.0
httpx>=0.23.0
